Separate packets by c2s/s2c
