+++ enigma-server/src/main/java/org/quiltmc/enigma/network/packet/s2c/EntryChangeS2CPacket.java
@@ -1,7 +1,10 @@
-package org.quiltmc.enigma.network.packet;
+package org.quiltmc.enigma.network.packet.s2c;
 
 import org.quiltmc.enigma.network.ClientPacketHandler;
 import org.quiltmc.enigma.api.translation.mapping.EntryChange;
+import org.quiltmc.enigma.network.packet.Packet;
+import org.quiltmc.enigma.network.packet.PacketHelper;
+import org.quiltmc.enigma.network.packet.c2s.ConfirmChangeC2SPacket;
 
 import java.io.DataInput;
 import java.io.DataOutput;
