+++ enigma-server/src/main/java/org/quiltmc/enigma/network/packet/s2c/SyncMappingsS2CPacket.java
@@ -1,4 +1,4 @@
-package org.quiltmc.enigma.network.packet;
+package org.quiltmc.enigma.network.packet.s2c;
 
 import org.quiltmc.enigma.api.source.TokenType;
 import org.quiltmc.enigma.network.ClientPacketHandler;
@@ -8,6 +8,9 @@ import org.quiltmc.enigma.api.translation.mapping.tree.EntryTree;
 import org.quiltmc.enigma.api.translation.mapping.tree.EntryTreeNode;
 import org.quiltmc.enigma.api.translation.mapping.tree.HashEntryTree;
 import org.quiltmc.enigma.api.translation.representation.entry.Entry;
+import org.quiltmc.enigma.network.packet.Packet;
+import org.quiltmc.enigma.network.packet.PacketHelper;
+import org.quiltmc.enigma.network.packet.c2s.ConfirmChangeC2SPacket;
 
 import java.io.DataInput;
 import java.io.DataOutput;
