+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/GuiController.java
@@ -34,8 +34,8 @@ import org.quiltmc.enigma.network.EnigmaServer;
 import org.quiltmc.enigma.gui.network.IntegratedEnigmaServer;
 import org.quiltmc.enigma.network.ServerMessage;
 import org.quiltmc.enigma.network.ServerPacketHandler;
-import org.quiltmc.enigma.network.packet.EntryChangeC2SPacket;
-import org.quiltmc.enigma.network.packet.LoginC2SPacket;
+import org.quiltmc.enigma.network.packet.c2s.EntryChangeC2SPacket;
+import org.quiltmc.enigma.network.packet.c2s.LoginC2SPacket;
 import org.quiltmc.enigma.network.packet.Packet;
 import org.quiltmc.enigma.api.source.DecompiledClassSource;
 import org.quiltmc.enigma.api.source.DecompilerService;
